Split bvec implementations from bvec.h into bvec.cpp
